Break and Continue Exercise
1.Create an HTML file named break_continue_js.html within the Codeup folder on your Mac.

2.Use Math.floor((Math.random()*50)+1) to generate a random number between 1 and 50. If the number is not odd, keep getting a new random number till it is odd.

3.Use console.log to log all the odd numbers from 1 to 50, except the random odd number you generated, by using break and continue. Try to match the output shown below (your random number will be different).
Finally, save your work and proceed to the next Lesson.
Your output should look like this:

Random odd number to skip is: 27
Here is an odd number: 1
Here is an odd number: 3
Here is an odd number: 5
Here is an odd number: 7
Here is an odd number: 9
Here is an odd number: 11
Here is an odd number: 13
Here is an odd number: 15
Here is an odd number: 17
Here is an odd number: 19
Here is an odd number: 21
Here is an odd number: 23
Here is an odd number: 25
Yikes! Skipping number: 27
Here is an odd number: 29
Here is an odd number: 31
Here is an odd number: 33
Here is an odd number: 35
Here is an odd number: 37
Here is an odd number: 39
Here is an odd number: 41
Here is an odd number: 43
Here is an odd number: 45
Here is an odd number: 47